pyproject.toml
[tool.poetry]
name = "mop-jenkins-bot-python"
version = "0.1.0"
description = ""
authors = ["Matt Carson <matt.carson224@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8.0"
disnake = "^2.5.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"